smtpserver: Fix handling of overlapping strings in reading of config file
This is the situation in cfgread.c, line 165 before the strcpy invocation:

  len = 5
  taillen = 25

  s0  0x7ffffffed327 "btreePE  $MAILVAR/db/smtp-policy"
  str 0x7ffffffed32e "  $MAILVAR/db/smtp-policy"

Then strcpy((char*)(s0+len), (const char *)str) is called and we get:

  s0 0x7ffffffed327 "btree  $MLVAR/db/smsmtp-policy"

Whops.  This happens because source and destination area are overlapping
but strcpy requires non-overlapping source and destination areas.

Unfortunately your fix is invalid because memcpy has the same restriction.
But you were close.  There is no strcpy variant that handles overlap
but memcpy's relative memmove does.  Also to ensure the new string is
zero terminated we copy one extra byte which is a zero byte.

  memmove(s0 + len, str, strlen((char*)str) + 1);

Signed-off-by: Ralf Baechle <ralf@linux-mips.org>